pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "DiegoSanchez",email = "diegosmtattoo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
python = ">=3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^5.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

